converted all domain model files to Kotlin

+++ core/src/main/java/org/mifos/mobilewallet/core/data/fineract/entity/mapper/AccountMapper.java
@@ -35,7 +35,7 @@ public class AccountMapper {
                 account.setId(savingAccount.getId());
                 account.setBalance(savingAccount.getAccountBalance());
                 account.setCurrency(currencyMapper.transform(savingAccount.getCurrency()));
-                account.setProductId(savingAccount.getProductId());
+                account.setProductId((long)savingAccount.getProductId());
 
                 accountList.add(account);
             }
